:recycle: move CLI classes to a separate module
